+++ app/src/main/java/org/oxycblt/auxio/playback/replaygain/ReplayGainPreAmp.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2022 Auxio Project
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+ 
+package org.oxycblt.auxio.playback.replaygain
+
+/** Represents the ReplayGain pre-amp values. */
+data class ReplayGainPreAmp(
+    /** The value to use when ReplayGain tags are present. */
+    val with: Float,
+    /** The value to use when ReplayGain tags are not present. */
+    val without: Float,
+)