+++ app/src/main/java/org/oxycblt/auxio/ui/ViewBindingDialogFragment.kt
@@ -17,7 +17,6 @@
  
 package org.oxycblt.auxio.ui
 
-import android.app.Dialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -35,17 +34,39 @@ import org.oxycblt.auxio.util.logD
 abstract class ViewBindingDialogFragment<T : ViewBinding> : DialogFragment() {
     private var _binding: T? = null
 
+    /**
+     * Inflate the binding from the given [inflater]. This should usually be done by the binding
+     * implementation's inflate function.
+     */
     protected abstract fun onCreateBinding(inflater: LayoutInflater): T
-    protected open fun onBindingCreated(binding: T, savedInstanceState: Bundle?) {}
-    protected open fun onDestroyBinding(binding: T) {}
+
+    /** Called during [onCreateDialog]. Dialog elements should be configured here. */
     protected open fun onConfigDialog(builder: AlertDialog.Builder) {}
 
+    /**
+     * Called during [onViewCreated] when the binding was successfully inflated and set as the view.
+     * This is where view setup should occur.
+     */
+    protected open fun onBindingCreated(binding: T, savedInstanceState: Bundle?) {}
+
+    /**
+     * Called during [onDestroyView] when the binding should be destroyed and all callbacks or
+     * leaking elements be released.
+     */
+    protected open fun onDestroyBinding(binding: T) {}
+
+    /** Maybe get the binding. This will be null outside of the fragment view lifecycle. */
     protected val binding: T?
         get() = _binding
 
+    /**
+     * Get the binding under the assumption that the fragment has a view at this state in the
+     * lifecycle. This will throw an exception if the fragment is not in a valid lifecycle.
+     */
     protected fun requireBinding(): T {
         return requireNotNull(_binding) {
-            "ViewBinding was not available, as the fragment was not in a valid state"
+            "ViewBinding was available. Fragment should be a valid state " +
+                "right now, but instead it was ${lifecycle.currentState}"
         }
     }
 
@@ -55,12 +76,11 @@ abstract class ViewBindingDialogFragment<T : ViewBinding> : DialogFragment() {
         savedInstanceState: Bundle?
     ): View = onCreateBinding(inflater).also { _binding = it }.root
 
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        return MaterialAlertDialogBuilder(requireActivity(), theme).run {
+    override fun onCreateDialog(savedInstanceState: Bundle?) =
+        MaterialAlertDialogBuilder(requireActivity(), theme).run {
             onConfigDialog(this)
             create()
         }
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -73,5 +93,6 @@ abstract class ViewBindingDialogFragment<T : ViewBinding> : DialogFragment() {
         super.onDestroyView()
         onDestroyBinding(requireBinding())
         _binding = null
+        logD("Fragment destroyed")
     }
 }